Header & Footer Progress. Removed Original React Page
Header is done.
Footer has top row and bottom row, but is missing middle elements such as the about, support, and more from kickstarter lists.
Footer has one list which must be changed to the about list. This was a test to print a list which did not work, but is in the code. The program is able to run.
I removed the elements of the original React page so only the header and footer show up.