implemented redirecting to CampaignOverview and initial data requests
only issue is that now i'm unsure how I should organize the data. It seems like there's a lot of information that should only be stored conditionally based on whether the campaign is published or not. It makes me think that perhaps I should re-organize the schemas and separate everything into unpublished and published campaigns.